+++ src/data-types/mailstream_compress.c
@@ -166,7 +166,7 @@ static ssize_t mailstream_low_compress_read(mailstream_low * s, void * buf, size
   do {
     /* if there is no compressed data, read more */
     if (strm->avail_in == 0) {
-      int read = data->ms->driver->mailstream_read(data->ms, data->input_buf, CHUNK_SIZE);
+      int read = (int) data->ms->driver->mailstream_read(data->ms, data->input_buf, CHUNK_SIZE);
       if (read <= 0) {
         return read;
       }
@@ -176,7 +176,7 @@ static ssize_t mailstream_low_compress_read(mailstream_low * s, void * buf, size
 
     /* set the output buffer */
     strm->next_out = buf;
-    strm->avail_out = count;
+    strm->avail_out = (int) count;
 
     /* uncompress any waiting data */
     zr = inflate(strm, Z_NO_FLUSH);
@@ -209,7 +209,7 @@ static ssize_t mailstream_low_compress_write(mailstream_low * s, const void * bu
 
   strm->next_in = (Bytef *)buf;
   /* we won't try to compress more than CHUNK_SIZE at a time so we always have enough buffer space */
-  int compress_len = MIN(count, CHUNK_SIZE);
+  int compress_len = MIN((int) count, CHUNK_SIZE);
   strm->avail_in = compress_len;
   strm->avail_out = CHUNK_SIZE;
   strm->next_out = data->output_buf;
