+++ src/driver/implementation/imap/imapdriver_cached.c
@@ -316,7 +316,7 @@ static int update_uid_cache(mailsession * session,
       }
       msg = carray_get(env_list->msg_tab, i);
       cache_item->uid = msg->msg_index;
-      cache_item->size = msg->msg_size;
+      cache_item->size = (uint32_t) msg->msg_size;
     
       carray_set(data->imap_uid_list, i, cache_item);
     }
@@ -934,7 +934,7 @@ static int boostrap_cache(mailsession * session)
     
     get_uid_from_filename(msg_uid);
     key.data = msg_uid;
-    key.len = strlen(msg_uid) + 1;
+    key.len = (unsigned int) strlen(msg_uid) + 1;
     value.data = NULL;
     value.len = 0;
     chash_set(keys_uid, &key, &value, NULL);
@@ -951,14 +951,14 @@ static int boostrap_cache(mailsession * session)
     chash_key(iter, &key);
     uid = key.data;
     
-    uidvalidity = strtoul(uid, &p1, 10);
+    uidvalidity = (uint32_t) strtoul(uid, &p1, 10);
     if (p1 == uid || * p1 != '-')
       continue;
     
     data->imap_uidvalidity = uidvalidity;
     
     p1++;
-    indx = strtoul(p1, &p2, 10);
+    indx = (uint32_t) strtoul(p1, &p2, 10);
     if (p2 == p1 || * p2 != '\0')
       continue;
     
@@ -1514,12 +1514,12 @@ static int imapdriver_cached_get_message_by_uid(mailsession * session,
   if (uid == NULL)
     return MAIL_ERROR_INVAL;
 
-  uidvalidity = strtoul(uid, &p1, 10);
+  uidvalidity = (uint32_t) strtoul(uid, &p1, 10);
   if (p1 == uid || * p1 != '-')
     return MAIL_ERROR_INVAL;
 
   p1++;
-  num = strtoul(p1, &p2, 10);
+  num = (uint32_t) strtoul(p1, &p2, 10);
   if (p2 == p1 || * p2 != '\0')
     return MAIL_ERROR_INVAL;
