+++ src/driver/implementation/mh/mhdriver_cached.c
@@ -287,7 +287,7 @@ static int write_max_uid_value(mailsession * session)
     goto free_mmapstr;
   }
 
-  r = fwrite(mmapstr->str, 1, mmapstr->len, f);
+  r = (int) fwrite(mmapstr->str, 1, mmapstr->len, f);
   if ((size_t) r != mmapstr->len) {
     res = MAIL_ERROR_FILE;
     goto free_mmapstr;
@@ -1281,7 +1281,7 @@ static int mhdriver_cached_get_message_by_uid(mailsession * session,
   if (uid == NULL)
     return MAIL_ERROR_INVAL;
 
-  indx = strtoul(uid, &p, 10);
+  indx = (uint32_t) strtoul(uid, &p, 10);
   if (p == uid || * p != '-')
     return MAIL_ERROR_INVAL;
