+++ src/driver/implementation/nntp/nntpdriver.c
@@ -1182,7 +1182,7 @@ static int nntpdriver_get_message_by_uid(mailsession * session,
   if (uid == NULL)
     return MAIL_ERROR_INVAL;
   
-  num = strtoul(uid, &p, 10);
+  num = (uint32_t) strtoul(uid, &p, 10);
   if ((p == uid) || (* p != '\0'))
     return MAIL_ERROR_INVAL;
